requirements.txt
requests
python-dotenv
pdfplumber
